move input parser to training script
