+++ train_reaction_graph_with_actions.py
@@ -1,11 +1,11 @@
 import os
 import gc
+import sys
 import datetime
-import pickle as pkl
+import argparse
 
 import numpy as np
-import pandas as pd
-from tqdm.autonotebook import tqdm
+import pickle as pkl
 
 import torch
 import torch.nn as nn
@@ -21,18 +21,19 @@ from sklearn.metrics import (
     f1_score,
 )
 
-from reaction_graph_actions.model import ReactionNet
-from reaction_graph_actions.action_rnn import LSTM
-from reaction_graph_actions.data import (
-    input_parser,
+from matgps.action_rnn import LSTM
+from matgps.reaction_graph_actions.model import ReactionNet
+from matgps.reaction_graph_actions.data import (
     ReactionData,
     collate_batch,
 )
-from reaction_graph_actions.utils import (
-    evaluate,
+from matgps.reaction_graph_actions.utils import (
+    evaluate
+)
+
+from matgps.utils import (
     save_checkpoint,
     load_previous_state,
-    cyclical_lr,
     LRFinder
 )
 
@@ -46,9 +47,11 @@ def get_class_weights(generator):
     # print(all_targets[:5])
     all_targets = torch.Tensor(all_targets)
     # print(all_targets.shape)
+
     # get weights of elements in batch
     num_elements = float(len(all_targets)) - (all_targets == 0).sum(dim=0)
     # print(num_elements)
+
     max_num_elements = torch.max(num_elements)
     max_num_elements_tensor = max_num_elements.repeat(len(num_elements))
     # print(max_num_elements_tensor)
@@ -131,10 +134,12 @@ def init_optim(model, weights=None):
         raise NameError("Only SGD or Adam is allowed as --optim")
 
     if args.clr:
-        clr = cyclical_lr(period=args.clr_period,
-                          cycle_mul=0.1,
-                          tune_mul=0.05,)
-        scheduler = optim.lr_scheduler.LambdaLR(optimizer, [clr])
+        scheduler = torch.optim.lr_scheduler.CyclicLR(
+            optimizer,
+            base_lr=args.learning_rate/10,
+            max_lr=args.learning_rate,
+            step_size_up=50,
+            cycle_momentum=False)
     else:
         scheduler = optim.lr_scheduler.MultiStepLR(optimizer, [])
 
@@ -154,8 +159,8 @@ def main():
 
     orig_prec_fea_len = dataset.prec_fea_dim
     orig_action_fea_len = dataset.action_fea_dim
-    print('orig precursor fea dim', orig_prec_fea_len)
-    print('action fea dim', orig_action_fea_len)
+    # print('orig precursor fea dim', orig_prec_fea_len)
+    # print('action fea dim', orig_action_fea_len)
 
     # load pretrained rnn
     rnn = LSTM(
@@ -193,7 +198,7 @@ def main():
     test_set = torch.utils.data.Subset(dataset, test_idx)
 
     # Ensure directory structure present
-    os.makedirs(f"models/", exist_ok=True)
+    os.makedirs("models/", exist_ok=True)
     os.makedirs("runs/", exist_ok=True)
     os.makedirs("results/", exist_ok=True)
 
@@ -334,14 +339,16 @@ def experiment(fold_id, run_id, args,
             model.to(args.device)
             criterion, optimizer, scheduler = init_optim(model)
 
+        best_loss = evaluate(
+            generator=val_generator,
+            model=model,
+            criterion=criterion,
+            optimizer=None,
+            device=args.device,
+            threshold=args.threshold,
+            task="val"
+        )
 
-        best_loss = evaluate(generator=val_generator,
-                                  model=model,
-                                  criterion=criterion,
-                                  optimizer=None,
-                                  device=args.device,
-                                  threshold=args.threshold,
-                                  task="val")
         start_epoch = 0
 
     # try except structure used to allow keyboard interupts to stop training
@@ -349,25 +356,29 @@ def experiment(fold_id, run_id, args,
     try:
         for epoch in range(start_epoch, start_epoch+args.epochs):
             # Training
-            t_loss = evaluate(generator=train_generator,
-                                             model=model,
-                                             criterion=criterion,
-                                             optimizer=optimizer,
-                                             device=args.device,
-                                             threshold=args.threshold,
-                                             task="train",
-                                             verbose=True)
+            t_loss = evaluate(
+                generator=train_generator,
+                model=model,
+                criterion=criterion,
+                optimizer=optimizer,
+                device=args.device,
+                threshold=args.threshold,
+                task="train",
+                verbose=True
+            )
 
             # Validation
             with torch.no_grad():
                 # evaluate on validation set
-                val_loss = evaluate(generator=val_generator,
-                                                       model=model,
-                                                       criterion=criterion,
-                                                       optimizer=None,
-                                                       device=args.device,
-                                                       threshold=args.threshold,
-                                                       task="val")
+                val_loss = evaluate(
+                    generator=val_generator,
+                    model=model,
+                    criterion=criterion,
+                    optimizer=None,
+                    device=args.device,
+                    threshold=args.threshold,
+                    task="val"
+                )
 
             # if epoch % args.print_freq == 0:
             print("Epoch: [{}/{}]\n"
@@ -447,13 +458,15 @@ def test_ensemble(fold_id, ensemble_folds, hold_out_set, fea_len, pretrained_rnn
         model.load_state_dict(checkpoint["state_dict"])
 
         model.eval()
-        idx, comp, pred, prec_embed, y_test, subset_accuracy, total = evaluate(generator=test_generator,
-                                            model=model,
-                                            criterion=criterion,
-                                            optimizer=None,
-                                            device=args.device,
-                                            threshold=args.threshold,
-                                            task="test")
+        idx, comp, pred, prec_embed, y_test, subset_accuracy, total = evaluate(
+            generator=test_generator,
+            model=model,
+            criterion=criterion,
+            optimizer=None,
+            device=args.device,
+            threshold=args.threshold,
+            task="test"
+        )
         y_ensemble[j,:] = pred
         y_ensemble_prec_embed[j,:] = prec_embed
 
@@ -478,7 +491,7 @@ def test_ensemble(fold_id, ensemble_folds, hold_out_set, fea_len, pretrained_rnn
     max_acc = max(subset_acc_dict.values())
     best_subset_acc = [{k:v} for k, v in subset_acc_dict.items() if v == max_acc]
     best_thresh = list(best_subset_acc[0].keys())[0]
-    #print(best_thresh)
+    # print(best_thresh)
     best_logit_thresh = np.log(best_thresh / (1 - best_thresh))
     # get uncertainty estimate from ensemble for best_subset_acc threshold
     ensemble_accs = [accuracy_score(test_elems, pred_logits > best_logit_thresh) for pred_logits in y_ensemble]
@@ -505,6 +518,249 @@ def test_ensemble(fold_id, ensemble_folds, hold_out_set, fea_len, pretrained_rnn
                                                    args.sample), 'wb') as f:
         pkl.dump(results, f)
     print(f'Dumped logits, targets, prec_embeddings, and ids to results file')
+
+
+def input_parser():
+    """
+    parse input
+    """
+    parser = argparse.ArgumentParser(
+        description="Inorganic Reaction Product Predictor, reaction graph model with actions")
+
+    # dataset inputs
+    parser.add_argument("--data-path",
+                        type=str,
+                        default="data/datasets/dataset_10_precs.pkl",
+                        metavar="PATH",
+                        help="dataset path")
+
+    parser.add_argument("--fea-path",
+                        type=str,
+                        default="data/embeddings/magpie_embed_10_precs.json",
+                        metavar="PATH",
+                        help="Precursor feature path")
+
+    parser.add_argument('--action-rnn',
+                        type=str,
+                        nargs='?',
+                        default='models/checkpoint_rnn_f-1_s-0_t-1.pth.tar',
+                        help="Path to trained action autoencoder")
+
+    parser.add_argument('--action-path',
+                        type=str,
+                        nargs='?',
+                        default='data/datasets/action_dict_10_precs.json',
+                        help="Path to action dictionary")
+
+    parser.add_argument('--elem-path',
+                        type=str,
+                        nargs='?',
+                        default='data/datasets/elem_dict_10_precs.json',
+                        help="Path to element dictionary")
+
+    parser.add_argument('--prec-type',
+                        type=str,
+                        nargs='?',
+                        default='magpie',
+                        help="Type of input, stoich or magpie")
+
+    parser.add_argument('--latent-dim',
+                        type=int,
+                        nargs='?',
+                        default=32,
+                        help='Latent dimension for RNN hidden state')
+
+    parser.add_argument('--intermediate-dim',
+                        type=int,
+                        nargs='?',
+                        default=256,
+                        help='Intermediate model dimension')
+
+    parser.add_argument('--target-dim',
+                        type=int,
+                        nargs='?',
+                        default=81,
+                        help='Target vector dimension')
+
+    parser.add_argument("--prec-fea-len",
+                        default=128,
+                        type=int,
+                        metavar="N",
+                        help="Dimension of node features")
+
+    parser.add_argument("--n-graph",
+                        default=5,
+                        type=int,
+                        metavar="N",
+                        help="number of graph layers")
+
+    parser.add_argument('--mask',
+                        action="store_true",
+                        default=False,
+                        help="Whether to mask output with precursor elements or not")
+
+    parser.add_argument('--amounts',
+                        action="store_true",
+                        default=False,
+                        help="use precursor amounts as weights")
+
+    parser.add_argument("--disable-cuda",
+                        action="store_true",
+                        help="Disable CUDA")
+
+    # restart inputs
+    parser.add_argument("--evaluate",
+                        action="store_true",
+                        help="skip network training stages checkpoint")
+
+    # dataloader inputs
+    parser.add_argument("--workers",
+                        default=0,
+                        type=int,
+                        metavar="N",
+                        help="number of data loading workers (default: 0)")
+
+    parser.add_argument("--batch-size", "--bsize",
+                        default=256,
+                        type=int,
+                        metavar="N",
+                        help="mini-batch size (default: 256)")
+
+    parser.add_argument("--val-size",
+                        default=0.0,
+                        type=float,
+                        metavar="N",
+                        help="proportion of data used for validation")
+
+    parser.add_argument("--test-size",
+                        default=0.2,
+                        type=float,
+                        metavar="N",
+                        help="proportion of data for testing")
+
+    parser.add_argument("--seed",
+                        default=0,
+                        type=int,
+                        metavar="N",
+                        help="seed for random number generator")
+
+    parser.add_argument("--sample",
+                        default=1,
+                        type=int,
+                        metavar="N",
+                        help="sub-sample the training set for learning curves")
+
+    # optimiser inputs
+    parser.add_argument("--epochs",
+                        default=60,
+                        type=int,
+                        metavar="N",
+                        help="number of total epochs to run")
+
+    parser.add_argument("--loss",
+                        default="BCE",
+                        type=str,
+                        metavar="str",
+                        help="choose a Loss Function")
+
+    parser.add_argument("--threshold",
+                        default=0.9,
+                        type=float,
+                        metavar='prob',
+                        help="Threshold for element presence in product (probability)")
+
+    parser.add_argument("--reg-weight",
+                        default=0,
+                        type=float,
+                        metavar="float",
+                        help="Weight for regularisation loss")
+
+    parser.add_argument("--optim",
+                        default="Adam",
+                        type=str,
+                        metavar="str",
+                        help="choose an optimizer; SGD, Adam or AdamW")
+
+    parser.add_argument("--learning-rate", "--lr",
+                        default=0.0001,
+                        type=float,
+                        metavar="float",
+                        help="initial learning rate (default: 3e-4)")
+
+    parser.add_argument("--momentum",
+                        default=0.9,
+                        type=float,
+                        metavar="float [0,1]",
+                        help="momentum (default: 0.9)")
+
+    parser.add_argument("--weight-decay",
+                        default=1e-6,
+                        type=float,
+                        metavar="float [0,1]",
+                        help="weight decay (default: 0)")
+
+    # ensemble inputs
+    parser.add_argument("--fold-id",
+                        default=1,
+                        type=int,
+                        metavar="N",
+                        help="identify the fold of the data")
+
+    parser.add_argument("--run-id",
+                        default=0,
+                        type=int,
+                        metavar="N",
+                        help="ensemble model id")
+
+    parser.add_argument("--ensemble",
+                        default=1,
+                        type=int,
+                        metavar="N",
+                        help="number ensemble repeats")
+
+    # transfer learning
+    parser.add_argument('--train-rnn',
+                        action="store_true",
+                        default=False,
+                        help="Train rnn for elem prediction as well")
+
+    parser.add_argument("--lr-search",
+                        action="store_true",
+                        help="perform a learning rate search")
+
+    parser.add_argument("--clr",
+                        default=True,
+                        type=bool,
+                        help="use a cyclical learning rate schedule")
+
+    parser.add_argument("--clr-period",
+                        default=100,
+                        type=int,
+                        help="how many learning rate cycles to perform")
+
+    parser.add_argument("--resume",
+                        action="store_true",
+                        help="resume from previous checkpoint")
+
+    parser.add_argument("--transfer",
+                        type=str,
+                        metavar="PATH",
+                        help="checkpoint path for transfer learning")
+
+    parser.add_argument("--fine-tune",
+                        type=str,
+                        metavar="PATH",
+                        help="checkpoint path for fine tuning")
+
+    args = parser.parse_args(sys.argv[1:])
+
+    if args.lr_search:
+        args.learning_rate = 1e-8
+
+    args.device = torch.device("cuda") if (not args.disable_cuda) and  \
+        torch.cuda.is_available() else torch.device("cpu")
+
+    return args
 
 if __name__ == "__main__":
     args = input_parser()
